+++ src/lib.rs
@@ -5,7 +5,26 @@
 //! Twitch types
 
 macro_rules! impl_extra {
+    (validated, $owned:path, $ref:path, $error:path) => {
+        impl<'a> TryFrom<&'a String> for &'a $ref {
+            type Error = $error;
+            fn try_from(string: &'a String) -> Result<Self, $error>  {
+                <$ref>::from_str(string.as_str())
+            }
+        }
+
+        impl_extra!(@all, $owned, $ref);
+    };
     ($owned:path, $ref:path) => {
+        impl<'a> From<&'a String> for &'a $ref {
+            fn from(string: &'a String) -> Self {
+                <$ref>::from_str(string.as_str())
+            }
+        }
+
+        impl_extra!(@all, $owned, $ref);
+    };
+    (@all, $owned:path, $ref:path) => {
         impl $ref {
             /// Get a
             #[doc = concat!("[`Cow<'_, ", stringify!($ref), ">`](std::borrow::Cow::Borrowed)")]
@@ -25,7 +44,72 @@ macro_rules! impl_extra {
                 ::std::borrow::Cow::Borrowed(&*owned)
             }
         }
+
+        impl<'a> crate::IntoCow<'a, $ref> for &'a $ref {
+            fn to_cow(self) -> ::std::borrow::Cow<'a, $ref> {
+                ::std::borrow::Cow::Borrowed(self)
+            }
+        }
+
+        impl<'a> crate::IntoCow<'a, $ref> for $owned {
+            fn to_cow(self) -> ::std::borrow::Cow<'a, $ref> {
+                ::std::borrow::Cow::Owned(self)
+            }
+        }
+
+        impl<'a> crate::IntoCow<'a, $ref> for &'a $owned {
+            fn to_cow(self) -> ::std::borrow::Cow<'a, $ref> {
+                ::std::borrow::Cow::Borrowed(self.as_ref())
+            }
+        }
     };
+}
+
+/// Convert a type into a [`Cow`](std::borrow::Cow)
+pub trait IntoCow<'a, Ref: ?Sized>
+where Ref: ToOwned {
+    /// Make the cow with proper ownership, muu
+    fn to_cow(self) -> std::borrow::Cow<'a, Ref>
+    where &'a Self: 'a;
+}
+
+impl<'a, R> IntoCow<'a, R> for std::borrow::Cow<'a, R>
+where
+    &'a R: Into<&'a R>,
+    R: ToOwned + ?Sized + 'a,
+    &'a R: Into<std::borrow::Cow<'a, R>>,
+    R::Owned: Into<std::borrow::Cow<'a, R>>,
+{
+    fn to_cow(self) -> std::borrow::Cow<'a, R> {
+        match self {
+            std::borrow::Cow::Borrowed(b) => b.into(),
+            std::borrow::Cow::Owned(o) => o.into(),
+        }
+    }
+}
+
+impl<'a, R> IntoCow<'a, R> for &'a str
+where
+    &'a str: Into<&'a R>,
+    R: ToOwned + ?Sized + 'a,
+{
+    fn to_cow(self) -> std::borrow::Cow<'a, R> { std::borrow::Cow::Borrowed(self.into()) }
+}
+
+impl<'a, R> IntoCow<'a, R> for &'a String
+where
+    &'a String: Into<&'a R>,
+    R: ToOwned + ?Sized + 'a,
+{
+    fn to_cow(self) -> std::borrow::Cow<'a, R> { std::borrow::Cow::Borrowed(self.into()) }
+}
+
+impl<'a, R> IntoCow<'a, R> for String
+where
+    String: Into<R::Owned>,
+    R: ToOwned + ?Sized + 'a,
+{
+    fn to_cow(self) -> std::borrow::Cow<'a, R> { std::borrow::Cow::Owned(self.into()) }
 }
 
 mod basic;
@@ -78,3 +162,35 @@ pub use crate::stream::*;
 pub use crate::time::*;
 #[cfg(feature = "user")]
 pub use crate::user::*;
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn lol() {
+        assert!(broadcaster_id("literal"));
+        assert!(!broadcaster_id(String::from("string")));
+        assert!(broadcaster_id(&String::from("ref string")));
+        assert!(broadcaster_id(UserIdRef::from_static("static ref")));
+        assert!(!broadcaster_id(UserId::new(String::from("owned"))));
+        assert!(broadcaster_id(&UserId::new(String::from("borrowed owned"))));
+        assert!(broadcaster_id(&*UserId::new(String::from("deref owned"))));
+        assert!(!broadcaster_id(std::borrow::Cow::Owned(UserId::new(
+            String::from("cow owned")
+        ))));
+        assert!(broadcaster_id(std::borrow::Cow::Borrowed(
+            UserIdRef::from_static("cow borrowed")
+        )));
+    }
+    /// aa
+    pub fn broadcaster_id<'a>(broadcaster_id: impl IntoCow<'a, UserIdRef> + 'a) -> bool {
+        struct K<'a> {
+            id: std::borrow::Cow<'a, UserIdRef>,
+        }
+        let k = K {
+            id: broadcaster_id.to_cow(),
+        };
+        matches!(k.id, std::borrow::Cow::Borrowed(_))
+    }
+}