+++ src/time.rs
@@ -2,7 +2,7 @@
 #[aliri_braid::braid(serde, validator, ord = "omit")]
 pub struct Timestamp;
 
-impl_extra!(Timestamp, TimestampRef);
+impl_extra!(validated, Timestamp, TimestampRef, TimestampParseError);
 
 impl aliri_braid::Validator for Timestamp {
     type Error = TimestampParseError;